Finish receiver
