+++ backend/HeatingDataMonitor.API/appsettings.json
@@ -1,27 +1,16 @@
 {
-    "Logging": {
-        "LogLevel": {
-            "Default": "Information",
-            "Microsoft": "Warning",
-            "Microsoft.Hosting.Lifetime": "Information"
-        }
-    },
-    "AllowedHosts": "*",
-    "ConnectionStrings": {
-        "HeatingDataDatabase": "Server=127.0.0.1;Port=5432;Database=heating_data_monitor;User Id=heatingDataMonitorUser;Password=dontworrythispasswordwillchangeinproduction;Max Auto Prepare=10;Auto Prepare Min Usages=2;"
-    },
-    "Serial": {
-        "PortName": "/dev/ttyAMA0",
-        "Delimiter": ";",
-        "BaudRate": 9600,
-        "DataBits": 8,
-        "Parity": "None",
-        "StopBits": "One",
-        "Handshake": "None",
-        "NewLine": "CRLF",
-        "Encoding": "us-ascii"
-    },
-    "Cache": {
-        "MaxSize": 100
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft": "Warning",
+      "Microsoft.Hosting.Lifetime": "Information"
     }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "HeatingDataDatabase": "Server=127.0.0.1;Port=5432;Database=heating_data_monitor;User Id=heatingDataMonitorUser;Password=dontworrythispasswordwillchangeinproduction;Max Auto Prepare=10;Auto Prepare Min Usages=2;"
+  },
+  "Cache": {
+    "MaxSize": 100
+  }
 }
