+++ backend/HeatingDataMonitor.Receiver/appsettings.Development.json
@@ -1,8 +1,9 @@
 {
   "Logging": {
     "LogLevel": {
-      "Default": "Information",
+      "Default": "Trace",
       "Microsoft.Hosting.Lifetime": "Information"
     }
-  }
+  },
+  "FakeSerialPortData": "sampleData.csv"
 }
