+++ backend/HeatingDataMonitor.Receiver/appsettings.json
@@ -5,6 +5,9 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
+  "ConnectionStrings": {
+    "HeatingDataDatabase": "Server=127.0.0.1;Port=5432;Database=heating_data_monitor;User Id=heatingDataMonitorUser;Password=dontworrythispasswordwillchangeinproduction;Max Auto Prepare=10;Auto Prepare Min Usages=2;"
+  },
   "Serial": {
     "PortName": "/dev/ttyAMA0",
     "BaudRate": 9600,
